More int -> long conversions for methods that take IDs as parameters.

+++ app/src/main/java/dnd/jon/spellbook/CharacterClassDao.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface CharacterClassDao extends DAO<CharacterClassEntry> {
 
     @Query("SELECT * FROM character_classes WHERE (character_id = :characterID AND class_id = :classID)")
-    CharacterClassEntry getEntryByIds(int characterID, int classID);
+    CharacterClassEntry getEntryByIds(long characterID, long classID);
 
     @Query("SELECT * FROM classes INNER JOIN (SELECT class_id FROM character_classes WHERE character_id = :characterID) AS res ON classes.id = res.class_id")
-    List<CasterClass> getVisibleClasses(int characterID);
+    List<CasterClass> getVisibleClasses(long characterID);
 
 }
